requirements.txt
imageio==2.9.0
keras==2.8.0
matplotlib==3.3.2
numpy==1.21.5
opencv_python==4.5.5.62
pandas==1.1.3
Pillow==9.2.0
streamlit==1.13.0
streamlit_option_menu==0.3.2
tensorflow==2.8.2
